+++ ccramic/app/entrypoint.py
@@ -9,14 +9,14 @@ from numba.core.errors import NumbaDeprecationWarning, NumbaPendingDeprecationWa
 import warnings
 
 _program = "ccramic"
-__version__ = "0.6.0"
+__version__ = "0.7.0"
 
 
 def get_current_dir():
     return str(os.path.abspath(os.path.join(os.path.dirname(__file__))))
 
 
-def init_app():
+def init_app(cli_config):
     # suppress numba depreciation warnings from umap
     warnings.simplefilter('ignore', category=NumbaDeprecationWarning)
     warnings.simplefilter('ignore', category=NumbaPendingDeprecationWarning)
@@ -76,7 +76,7 @@ def init_app():
         from .app import init_dashboard
         # use a unique uuid for the session id I/O
         authentic_user = str(uuid.uuid1())
-        app = init_dashboard(app, authentic_user)
+        app = init_dashboard(app, authentic_user, config=cli_config)
         # init_callbacks(app)
 
         return app